move pre_promote to separate class
